CSN: CSN-asimcore-379
Sync with Intel rev-3413

Author:	mgambhir
Last modification:	Rev 3413 - 2011-11-15 18:57:32
Log message:	CSN: CSN-simcore-3413

- Patch by Baida, Yuriy V to fix warnings when built with gcc 4.6

+++ lib/libdraldb/include/asim/TagVecDenseItemIdx.h
@@ -142,7 +142,7 @@ TagVecDenseItemIdx::getTagValue(INT32 cycle, UINT64* value, UINT32* atcycle)
     INT32 i=0;
     bool fnd = false;
 	UINT32 ccycle = 0;
-    while (!fnd && ((valvec[i].offset+baseCycle)<=cycle) && (i<nextEntry))
+    while (!fnd && ((valvec[i].offset+baseCycle)<=(UINT32)cycle) && (i<nextEntry))
     {
 		ccycle = valvec[i].offset+baseCycle;
         fnd = ccycle == (UINT32)cycle ;
